fix: User PK 타입 수정
- INT -> BIGINT (카카오 회원번호 타입 반영)

+++ models/event.ts
@@ -25,7 +25,7 @@ class EventModel extends Model {
   promisingId: number;
 
   @ForeignKey(() => User)
-  @Column({ type: DataType.INTEGER })
+  @Column({ type: DataType.BIGINT })
   userId: number;
 
   @BelongsTo(() => PromisingModel, 'promisingId')
